Change the schema to something closer to the original request
Rather than expecting arrays of candidates in preference order, assume
the input is a set of rows with (candidate,preference) for
each (specified) candidate for each voter in each election. The bulk
of the query is unchanged since the array format is far easier to work
with, we just compute it first.

Also change to using explicit PREPARE for convenience, and enlarge the
test data a bit.