Fix the logic.
Reading the spec more closely, the logic allows the election of
multiple candidates in a single round, but only one candidate's
ballots are to be transferred in any round. That means we need to keep
track of pending transfers for candidates already elected.

To help in verification, add annotations to the output so we can see
the intermediate round totals and actions.

Include an actual (anonymized) election in the test data set to have a
known result to compare with.